requirements.txt
numpy
scikit-learn
pygame
pyglet
colorama
pyyaml
wandb

--extra-index-url https://download.pytorch.org/whl/cu126
torch==2.6.0+cu126